requirements.txt
boto3
awscli 
numpy
faiss-cpu
streamlit